# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Construction Extracts Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{userId}/approve:
    post:
      operationId: approveUser
      tags: [users]
      summary: Approve a user registration (admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}/reject:
    post:
      operationId: rejectUser
      tags: [users]
      summary: Reject a user registration (admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, supervisor, contract_supervisor, user]
        contractCompany:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        company:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        hospital:
          type: string
          nullable: true
        jobTitle:
          type: string
          nullable: true
        contractNumber:
          type: string
          nullable: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        lastPage:
          type: string
          nullable: true
        lastPageAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - role
        - status
        - createdAt
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
...