# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Construction Extracts Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileRole } from "./userProfileRole";
import type { UserProfileStatus } from "./userProfileStatus";

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  name?: string | null;
  role: UserProfileRole;
  contractCompany?: string | null;
  status: UserProfileStatus;
  company?: string | null;
  phone?: string | null;
  hospital?: string | null;
  jobTitle?: string | null;
  contractNumber?: string | null;
  lastLoginAt?: Date | null;
  lastPage?: string | null;
  lastPageAt?: Date | null;
  createdAt: Date;
}